# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A Task queue backed by Elasticsearch",
        "title": "Tasques API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/recurring_tasques": {
            "get": {
                "description": "Lists persisted Recurring Tasks",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recurring-tasks"
                ],
                "summary": "List Recurring Tasks",
                "operationId": "list-existing-recurring-tasks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/recurring.Task"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new Recurring Task",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recurring-tasks"
                ],
                "summary": "Add a new Recurring Task",
                "operationId": "create-recurring-task",
                "parameters": [
                    {
                        "description": "The request body",
                        "name": "newRecurringTask",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/recurring.NewTask"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/recurring.Task"
                        }
                    },
                    "400": {
                        "description": "Invalid JSON",
                        "schema": {
                            "$ref": "#/definitions/common.Body"
                        }
                    },
                    "409": {
                        "description": "Id in use",
                        "schema": {
                            "$ref": "#/definitions/common.Body"
                        }
                    }
                }
            }
        },
        "/recurring_tasques/{id}": {
            "get": {
                "description": "Retrieves a persisted Recurring Task",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recurring-tasks"
                ],
                "summary": "Get a Recurring Task",
                "operationId": "get-existing-recurring-task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The id of the Recurring Task",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/recurring.Task"
                        }
                    },
                    "404": {
                        "description": "Recurring Task does not exist",
                        "schema": {
                            "$ref": "#/definitions/common.Body"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates a persisted Recurring Task",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recurring-tasks"
                ],
                "summary": "Update a Recurring Task",
                "operationId": "update-existing-recurring-task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The id of the Recurring Task",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The request body",
                        "name": "recurringTaskUpdate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/recurring.TaskUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/recurring.Task"
                        }
                    },
                    "404": {
                        "description": "Recurring Task does not exist",
                        "schema": {
                            "$ref": "#/definitions/common.Body"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a persisted Recurring Task",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recurring-tasks"
                ],
                "summary": "Delete a Recurring Task",
                "operationId": "delete-existing-recurring-task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The id of the Recurring Task",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/recurring.Task"
                        }
                    },
                    "404": {
                        "description": "Recurring Task does not exist",
                        "schema": {
                            "$ref": "#/definitions/common.Body"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.Body": {
            "type": "object",
            "required": [
                "message"
            ],
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Something went wrong :("
                }
            }
        },
        "common.Metadata": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "modified_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "version": {
                    "type": "object",
                    "$ref": "#/definitions/common.Version"
                }
            }
        },
        "common.Version": {
            "type": "object",
            "properties": {
                "primary_term": {
                    "type": "integer"
                },
                "seq_num": {
                    "type": "integer"
                }
            }
        },
        "recurring.NewTask": {
            "type": "object",
            "required": [
                "id",
                "schedule_expression",
                "task_definition"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "schedule_expression": {
                    "description": "A schedule expression; can be any valid cron expression, with some support for simple macros",
                    "type": "string",
                    "example": "@every 1m"
                },
                "task_definition": {
                    "type": "object",
                    "$ref": "#/definitions/recurring.TaskDefinition"
                }
            }
        },
        "recurring.Task": {
            "type": "object",
            "required": [
                "id",
                "metadata",
                "schedule_expression",
                "task_definition"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "loaded_at": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "$ref": "#/definitions/common.Metadata"
                },
                "schedule_expression": {
                    "description": "A schedule expression; can be any valid cron expression, with some support for simple macros",
                    "type": "string",
                    "example": "@every 1m"
                },
                "task_definition": {
                    "type": "object",
                    "$ref": "#/definitions/recurring.TaskDefinition"
                }
            }
        },
        "recurring.TaskDefinition": {
            "type": "object",
            "required": [
                "kind",
                "queue"
            ],
            "properties": {
                "args": {
                    "type": "object"
                },
                "context": {
                    "type": "object"
                },
                "kind": {
                    "type": "string",
                    "example": "sayHello"
                },
                "priority": {
                    "type": "integer"
                },
                "processing_timeout": {
                    "type": "string",
                    "example": "30m"
                },
                "queue": {
                    "type": "string",
                    "example": "run-later"
                },
                "retry_times": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
        "recurring.TaskUpdate": {
            "type": "object",
            "properties": {
                "schedule_expression": {
                    "description": "A schedule expression; can be any valid cron expression, with some support for simple macros",
                    "type": "string",
                    "example": "@every 1m"
                },
                "task_definition": {
                    "type": "object",
                    "$ref": "#/definitions/recurring.TaskDefinition"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}